+++ src/previewdef/worldmap/loader/terrain.ts
@@ -1,7 +1,9 @@
 import { CustomMap, DetailValue, Enum, SchemaDef } from "../../../hoiformat/schema";
-import { FileLoader, convertColor, LoadResultOD } from "./common";
-import { Terrain } from "../definitions";
+import { FileLoader, convertColor, LoadResultOD, FolderLoader, mergeInLoadResult } from "./common";
+import { MapLoaderExtra, Terrain } from "../definitions";
 import { readFileFromModOrHOI4AsJson } from "../../../util/fileloader";
+import { LoadResult, LoaderSession } from '../../../util/loader/loader';
+import { localize } from '../../../util/i18n';
 
 interface TerrainFile {
     categories: CustomMap<TerrainCategory>
@@ -25,11 +27,42 @@ const terrainFileSchema: SchemaDef<TerrainFile> = {
     },
 };
 
-export class TerrainDefinitionLoader extends FileLoader<Terrain[]> {
+export class TerrainDefinitionLoader extends FolderLoader<Terrain[], Terrain[]> {
     constructor() {
-        super('common/terrain/00_terrain.txt');
+        super('common/terrain', TerrainFileLoader);
+    }
+    
+    protected mergeFiles(fileResults: LoadResult<Terrain[], MapLoaderExtra>[], session: LoaderSession): Promise<LoadResult<Terrain[], MapLoaderExtra>> {
+        const results =  mergeInLoadResult(fileResults, 'result');
+        const terrainMap: Record<string, Terrain> = {};
+        const warnings = mergeInLoadResult(fileResults, 'warnings');
+
+        for (const terrain of results) {
+            if (terrain.name in terrainMap) {
+                warnings.push({
+                    source: [],
+                    text: localize('worldmap.warnings.terraindefinedtwice', 'Terrain {0} is defined in two files: {1}, {2}.',
+                        terrain.name, terrain.file, terrainMap[terrain.name].file),
+                    relatedFiles: [terrain.file, terrainMap[terrain.name].file],
+                });
+            } else {
+                terrainMap[terrain.name] = terrain;
+            }
+        }
+
+        return Promise.resolve({
+            result: Object.values(terrainMap),
+            warnings,
+            dependencies: [this.folder + '/*'],
+        });
     }
 
+    public toString() {
+        return `[TerrainDefinitionLoader]`;
+    }
+}
+
+export class TerrainFileLoader extends FileLoader<Terrain[]> {
     protected async loadFromFile(): Promise<LoadResultOD<Terrain[]>> {
         return {
             result: await loadTerrains(this.file),
@@ -38,7 +71,7 @@ export class TerrainDefinitionLoader extends FileLoader<Terrain[]> {
     }
 
     public toString() {
-        return `[TerrainDefinitionLoader]`;
+        return `[TerrainFileLoader ${this.file}]`;
     }
 }
 
@@ -48,6 +81,6 @@ async function loadTerrains(file: string): Promise<Terrain[]> {
         const name = v._key;
         const color = convertColor(v._value.color);
         const isNaval = v._value.naval_terrain ?? false;
-        return { name, color, isNaval };
+        return { name, color, isNaval, file };
     });
 }